requirements.txt
playwright==1.37.0
python-dotenv==1.0.0
openai==0.27.8
schedule==1.2.0
rich==13.4.2
beautifulsoup4==4.12.2
lxml==4.9.3
